requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
celery==5.3.4
redis==5.0.1
flower==2.0.0
faiss-cpu 
sentence-transformers 
pandas
langchain 
langchain-anthropic
python-dotenv  
fuzzywuzzy
pydantic-settings 
